requirements.txt
astroid==2.2.5
autopep8==1.4.4
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
Django==2.2.3
djangorestframework==3.10.2
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
psycopg2==2.8.3
psycopg2-binary==2.8.3
pycodestyle==2.5.0
pylint==2.3.1
pylint-django==2.0.11
pylint-plugin-utils==0.5
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.2
